Updated DailyMenu repositories and models

+++ Exebite.DataAccess/Repositories/DailyMenu/IDailyMenuCommandRepository.cs
@@ -1,6 +1,6 @@
 ﻿namespace Exebite.DataAccess.Repositories
 {
-    public interface IDailyMenuCommandRepository : IDatabaseCommandRepository<int, DailyMenuInsertModel, DailyMenuUpdateModel>
+    public interface IDailyMenuCommandRepository : IDatabaseCommandRepository<long, DailyMenuInsertModel, DailyMenuUpdateModel>
     {
     }
 }
